feat: rename cartesian product function
