+++ src/shared/lib/helpers/getCartesianProduct.ts
